requirements.txt
streamlit
openai
chromadb
openai-whisper
python-dotenv
faiss-cpu
tiktoken
os